requirements.txt
faker
faker-stalker-names
discord.py
python-dotenv